fw-4383, merge
