+++ firstvoices/backend/models/media.py
@@ -7,6 +7,7 @@ from django.db import models
 from django.utils.translation import gettext as _
 from PIL import Image as PILImage
 
+import backend.permissions.predicates.edit
 from backend.permissions import predicates
 
 from .base import AudienceMixin, BaseSiteContentModel
@@ -20,9 +21,9 @@ class Person(BaseSiteContentModel):
         verbose_name_plural = _("People")
         rules_permissions = {
             "view": predicates.has_visible_site,
-            "add": predicates.can_add_core_uncontrolled_data,
-            "change": predicates.can_edit_core_uncontrolled_data,
-            "delete": predicates.can_edit_core_uncontrolled_data,
+            "add": backend.permissions.predicates.edit.can_add_core_uncontrolled_data,
+            "change": backend.permissions.predicates.edit.can_edit_core_uncontrolled_data,
+            "delete": backend.permissions.predicates.edit.can_edit_core_uncontrolled_data,
         }
 
     # from dc:title
