+++ firstvoices/backend/permissions/predicates/__init__.py
@@ -1,5 +1,4 @@
 from .base import *  # noqa F401
 from .edit import *  # noqa F401
-from .edit2 import *  # noqa F401
 from .view import *  # noqa F401
 from .view_models import *  # noqa F401
