--- firstvoices/backend/permissions/predicates/edit2.py
@@ -1,40 +1,0 @@
-from rules import predicate
-
-from . import base
-
-
-@predicate
-def can_add_core_uncontrolled_data(user, obj):
-    """
-    Same as ``can_edit_core_uncontrolled_data`` but does not check the visibility of the object that is saved in the db.
-
-    - Assistant has permission for Team-visibility data
-    - Editor, Language Admin, and Superadmin have permission
-    """
-    return (
-        base.has_at_least_editor_membership(user, obj)
-        | base.is_superadmin(user, obj)
-        | (
-            base.has_at_least_assistant_membership(user, obj)
-            & base.is_team_obj(user, obj)  # will be called with a site object
-        )
-    )
-
-
-@predicate
-def can_edit_core_uncontrolled_data(user, obj):
-    """
-    Same as ``can_add_core_uncontrolled_data`` but also checks the visibility of the object that is saved in the db.
-
-    - Assistant has permission for Team-visibility data
-    - Editor, Language Admin, and Superadmin have permission
-    """
-    return (
-        base.has_at_least_editor_membership(user, obj)
-        | base.is_superadmin(user, obj)
-        | (
-            base.has_at_least_assistant_membership(user, obj)
-            & base.has_team_site(user, obj)
-            & base.has_saved_team_site(user, obj)
-        )
-    )
