+++ firstvoices/backend/tests/test_permissions/__init__.py
@@ -1,7 +1,6 @@
 from . import (  # noqa F401
     test_permission_utils,
     test_predicate_base,
-    test_predicate_edit2,
     test_predicate_view,
     test_predicate_view_models,
 )
